requirements.txt
python-dotenv
supabase
openai
colorama
pydantic
numpy